Update CommandProcessor.
Add `go_to_command_in_collection` as helper for commands that need to be in other collection,
supposed to work for commands that requires to go to a branch or a different collection.

Changes `go_to_command(int, Collection)` to `go_to_command(int)` to move in the same `main_collection`.

Modifies the way `get_next_command_position` works, relying on `CommandCollection.get_next_command_position_according()`

Modifies `go_to_next_command` to make sure to verify if possible to go to the next command before trying to move.