Update Editor
Add `disable` and `enable`  methods to notify the editor and all childs that editor was enabled or disabled.
Those methods are called on very specific scenarios and can be or not related to editing of a collection.

Add `EditorConstants` (constants.gd) to keep track of editor constants that are related to editor but used by other non main editor related scripts.